implement document with have
value is provided to @@have before the mustache is executed, template
rendering is a lightweight process that does not involve any fetching

this will not work so well when we want to fetch values of all of the
undocumented fields.  we were going to focus on editing documentation
though, which means we really need to retrieve the current values of
'notes' and 'status' at least from MySQL, display them in a form, and
react appropriately when someone submits an edit (save it, goto next)

these fields are actually all documented now, according to "those with
status or notes are documented" standard which was established to find
this list from some cherry-picked tables that were thought to be most
important for documentation purposes.  form should allow "review mode"
but not by default. (so, meet the standard before setting another one)