Pass #2 notation
The fieldedit and flist interfaces have been updated to obey the #2
rule, that records whose status codes do not end in #2 are not done.

The only place where status codes are displayed and we don't chomp
(when present) the #2 before displaying status is in :flist, where
entries that end in #2 are already excluded from display anyway as
"done".

A future version might change back to "done, unless notes is blank"
That would make a convenient way to reject bad notes... by deleting
the contents of the notes field.

If they were backed up somewhere...
We are not done with Pass #2 yet.